refactor: 학수번호 타입 수정
- Integer -> Long
- Wrapper 타입을 쓴 이유는 존재하지 않을 경우 null 값을 허용하기 위함

+++ src/main/java/com/gonghak98/v2/common/DevDataLoader.java
@@ -81,7 +81,7 @@ public class DevDataLoader implements ApplicationRunner {
                     while ((line = br.readLine()) != null) {
                         final String[] split = line.split(DELIMITER);
                         try {
-                            Integer courseId = Integer.parseInt(split[0]);
+                            Long courseId = Long.parseLong(split[0]);
                             String name = split[1];
                             Double point = Double.parseDouble(split[2]);
