+++ src/test/java/com/gonghak98/v2/abeek/basic/msc/MscBasicTest.java
@@ -3,10 +3,10 @@ package com.gonghak98.v2.abeek.basic.msc;
 import static com.gonghak98.v2.abeek.fixture.BasicFixture.createMscBasic;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.gonghak98.v2.abeek.fixture.GivenObjectFixture;
 import com.gonghak98.v2.report.domain.abeek.AreaType;
 import com.gonghak98.v2.report.domain.abeek.basic.Basic;
 import com.gonghak98.v2.report.domain.abeek.dto.RequirementResult;
-import com.gonghak98.v2.abeek.fixture.GivenObjectFixture;
 import com.gonghak98.v2.report.domain.student.CompletedCourse;
 import java.util.List;
 import java.util.stream.Stream;
@@ -53,19 +53,19 @@ class MscBasicTest {
 
         private static Stream<Arguments> provideAllEssentialMscCourses() {
             List<CompletedCourse> essentialCourses = List.of(
-                CompletedCourse.builder().id(11300).name("고급프로그래밍활용").build(),
-                CompletedCourse.builder().id(7330).name("확률및통계").build(),
-                CompletedCourse.builder().id(9912).name("C프로그래밍및실습").build(),
-                CompletedCourse.builder().id(1357).name("미적분학1").build(),
-                CompletedCourse.builder().id(304).name("공업수학1").build(),
-                CompletedCourse.builder().id(9913).name("고급C프로그래밍및실습").build(),
-                CompletedCourse.builder().id(1725).name("선형대수").build(),
-                CompletedCourse.builder().id(11320).name("인공지능과빅데이터").build(),
-                CompletedCourse.builder().id(11678).name("기초전자물리").build()
+                CompletedCourse.builder().id(11300L).name("고급프로그래밍활용").build(),
+                CompletedCourse.builder().id(7330L).name("확률및통계").build(),
+                CompletedCourse.builder().id(9912L).name("C프로그래밍및실습").build(),
+                CompletedCourse.builder().id(1357L).name("미적분학1").build(),
+                CompletedCourse.builder().id(304L).name("공업수학1").build(),
+                CompletedCourse.builder().id(9913L).name("고급C프로그래밍및실습").build(),
+                CompletedCourse.builder().id(1725L).name("선형대수").build(),
+                CompletedCourse.builder().id(11320L).name("인공지능과빅데이터").build(),
+                CompletedCourse.builder().id(11678L).name("기초전자물리").build()
             );
 
             List<CompletedCourse> otherCourses = List.of(
-                CompletedCourse.builder().id(Integer.MAX_VALUE).name("테스트과목").build()
+                CompletedCourse.builder().id(Long.MAX_VALUE).name("테스트과목").build()
             );
 
             return Stream.of(
@@ -75,15 +75,15 @@ class MscBasicTest {
 
         private static Stream<Arguments> provideNotAllEssentialMscCourses() {
             List<CompletedCourse> essentialCourses = List.of(
-                CompletedCourse.builder().id(11300).name("고급프로그래밍활용").build(),
-                CompletedCourse.builder().id(7330).name("확률및통계").build(),
-                CompletedCourse.builder().id(9912).name("C프로그래밍및실습").build(),
-                CompletedCourse.builder().id(1357).name("미적분학1").build(),
-                CompletedCourse.builder().id(304).name("공업수학1").build()
+                CompletedCourse.builder().id(11300L).name("고급프로그래밍활용").build(),
+                CompletedCourse.builder().id(7330L).name("확률및통계").build(),
+                CompletedCourse.builder().id(9912L).name("C프로그래밍및실습").build(),
+                CompletedCourse.builder().id(1357L).name("미적분학1").build(),
+                CompletedCourse.builder().id(304L).name("공업수학1").build()
             );
 
             List<CompletedCourse> otherCourses = List.of(
-                CompletedCourse.builder().id(Integer.MAX_VALUE).name("테스트과목").build()
+                CompletedCourse.builder().id(Long.MAX_VALUE).name("테스트과목").build()
             );
 
             return Stream.of(
